Phase 10 completed - Modding support, proven by a variant with zero edits to core
The phase the architecture exists for. A rules variant now ships as a directory,
and the claim is tested rather than asserted.

New: src/here_to_slay/modding/
  * plugins.py  - a pack's plugin.py, imported and applied. One Plugin
    declaration reaches BOTH the engine registries and the validator
    vocabulary, so an op cannot run-but-fail-validation or validate-but-explode.
    Registration is deferred rather than done at import time, because
    sys.modules caches a module and its decorators would fire exactly once per
    process.
  * scaffold.py - hts new-pack <name> [--plugin], writing a pack that validates
    and deals immediately, commented where the merge rules bite.
  * diffing.py  - hts diff-pack base variant, comparing the resolved tables the
    engine really walks, with id-keyed paths so adding one action does not
    report every later action as edited.

modding/ is a fourth package on purpose: importing plugin.py needs the pack
directory (core may not read it) and the engine registries (content may not
import them). test_layering.py now asserts the new line.

Sample variant: data/variants/overclock
  A seventh class (hacker), a per-seat cache zone, three actions
  (upload/download/overclock), a reaction window on an event the engine has
  never heard of, a win condition that replaces full_party, one op in each of
  the five registries, and a dotted patch to a base Monster - with zero edits
  to core/. TestNothingInCoreKnowsAboutThisPack greps every engine source file
  and asks a subprocess what a fresh interpreter has registered; both come back
  empty.

Engine and tooling fixes the variant found:
  * a required pack one directory up is now discovered without --search-path,
    so hts validate data/variants/x just works
  * one plugin file reached by two path spellings built two Plugin objects and
    collided on install; paths are resolved before the module cache is keyed
  * diff-pack diffed the global registries, which reports nothing once a pack
    is already installed; it reads the Plugin objects instead

UI, same theme, no visual change to the base game:
  * the class tracker counted len(CLASS_COLOURS) - a palette - instead of
    rules.classes, so a seven-class pack showed 6/6 and dropped a dot. Party
    row, opponent strips and the rules screen read the loaded rule set now.

Also: docs/modding_guide.md (the tour), Phase 10 marked done in build_plan.md
with its decisions, and the ruff debt the Phase 9 UI rework left behind is
cleared - ruff check . is clean across the repository again.

uv run pytest -> 940 passed, and again under HTS_STRICT=1.
uv run hts validate data/variants/overclock --strict -> OK, 95 definitions.

+++ src/here_to_slay/ui/pygame/layout.py
@@ -24,10 +24,10 @@ NARROW = 1180
 
 def _lerp_rect(a: pygame.Rect, b: pygame.Rect, t: float) -> pygame.Rect:
     return pygame.Rect(
-        int(round(a.x + (b.x - a.x) * t)),
-        int(round(a.y + (b.y - a.y) * t)),
-        max(0, int(round(a.w + (b.w - a.w) * t))),
-        max(0, int(round(a.h + (b.h - a.h) * t))),
+        round(a.x + (b.x - a.x) * t),
+        round(a.y + (b.y - a.y) * t),
+        max(0, round(a.w + (b.w - a.w) * t)),
+        max(0, round(a.h + (b.h - a.h) * t)),
     )
 
 
@@ -269,8 +269,10 @@ class LayoutManager:
         """
         w, h = self.width, self.height
         # Left stays compact; right is wider so action labels stay readable.
-        left_w = max(200, min(280 if not compact else 220, int(w * (0.14 if not compact else 0.12))))
-        right_w = max(300, min(420 if not compact else 340, int(w * (0.22 if not compact else 0.18))))
+        left_w = max(200, min(280 if not compact else 220,
+                              int(w * (0.14 if not compact else 0.12))))
+        right_w = max(300, min(420 if not compact else 340,
+                               int(w * (0.22 if not compact else 0.18))))
         pad = max(8, gap)
 
         # --- left column: dice at bottom, merged journal above -----------
@@ -498,7 +500,7 @@ class LayoutManager:
                 setattr(self, name, _lerp_rect(a, b, t))
         for name in self._LERP_INTS:
             if name in frm and name in to:
-                setattr(self, name, int(round(frm[name] + (to[name] - frm[name]) * t)))
+                setattr(self, name, round(frm[name] + (to[name] - frm[name]) * t))
         seats_a = frm.get("seats") or []
         seats_b = to.get("seats") or []
         for i, seat in enumerate(self.seats):
@@ -508,10 +510,10 @@ class LayoutManager:
             seat.rect = _lerp_rect(a["rect"], b["rect"], t)
             seat.party_rect = _lerp_rect(a["party_rect"], b["party_rect"], t)
             seat.centre = (
-                int(round(a["centre"][0] + (b["centre"][0] - a["centre"][0]) * t)),
-                int(round(a["centre"][1] + (b["centre"][1] - a["centre"][1]) * t)),
+                round(a["centre"][0] + (b["centre"][0] - a["centre"][0]) * t),
+                round(a["centre"][1] + (b["centre"][1] - a["centre"][1]) * t),
             )
-            seat.card_w = int(round(a["card_w"] + (b["card_w"] - a["card_w"]) * t))
+            seat.card_w = round(a["card_w"] + (b["card_w"] - a["card_w"]) * t)
             seat.angle = a["angle"] + (b["angle"] - a["angle"]) * t
 
     @property
@@ -631,7 +633,7 @@ class LayoutManager:
                 party = pygame.Rect(self.party_rect)
                 card_w = self.party_card_w
             else:
-                # Side/north wedges — larger than the old 34–56px rail minis.
+                # Side/north wedges — larger than the old 34-56px rail minis.
                 card_w = max(64, min(96, int(min(table.width, table.height) * 0.07)))
                 self.seat_card_w = card_w
                 self.rail_card_w = card_w
